Loan Approval/deployment/requirements.txt
streamlit
pandas
scikit-learn
seaborn
joblib
numpy
imbalanced-learn
plotly.express
xgboost
